pytest: don’t depend on near-sdk in empty-contract-rs
There’s very little reason to make the empty-contract-rs depend on
near-sdk.  The code is trivial and only uses the log method which
can be easily called directly without the near-sdk wrapper.

This also fixes the issue where the test was not reproducible as the
empty-contract-rs would depend on the near-sdk from the git
repository rather than a released crate.

An obvious question which arises is why not use near-test-contracts
like the rest of pytests.  That would be an ideal endgoal but it
doesn’t interact well with how NayDuck runs mocknet tests.
Specifically, near-test-contracts needs to be built and NayDuck
currently skips build stage when running mocknet tests.  Long term
this could be addressed but for now this change improves the
situation so it’s what we’ve got.

Fixes: https://github.com/near/nearcore/issues/4480