Makefile: add `neard` target which just buids neard binary
The new target will be mentioned in the ‘Compile and Run without
Container’ document† and replace `cargo build` invocations listed
there.

This is because the plan is to disable LTO in the release build
profile and thus to get a best optimised version of the binary
environment variables need to be set.  The Makefile target does that.

† https://docs.near.org/docs/develop/node/validator/compile-and-run-a-node

Issue: https://github.com/near/nearcore/issues/4403