refactor: drop unneeded `derive(Serialize)`
Serialize is a big promise, so better not to make it if we don't have
to.

Additionally, structs which are both `derive(Serialize, BorshSerialize)`
are suuuper suspicious and a code smell. bors and JSON (what we use
Serialize for) have very different format compatibility guarantees,
keeping a type backwards compatible with both formats at the same time
is tough.