fix(runtime): Fix validator stake host function (#4535)
In #3821, we correctly fixed the meaning of `last_block_hash` in `ApplyState`. However, this unveiled a bug that was previously masked: we used `last_block_hash` in the implementation of `validator_*` function and unfortunately, when the chunk is applied, the block is not yet added to epoch manager and therefore trying to access `last_block_hash` leads to a `EpochError`, which in turn leads to the chunk getting rejected. This is what caused the testnet to stall on June 29th. This fix has already been deployed on both testnet and mainnet. This PR ports the fix to master.

Test plan
----------
* `test_validator_stake_host_function`