pytest: add a pre-commit check testing if all pytests are enabled
Add the scripts/check_pytest.py to the ‘sanity checks’ pre-commit
tests to make sure that all pytest tests are enabled in one of the
nightly/*.txt files.

Because at the moment there are nine files which do not meet that
criteria, add them to nightly/complete_nayduck_set.txt with a comment
documenting the state.

Issue: https://github.com/near/nearcore/issues/4552
Fixes: https://github.com/near/nearcore/issues/2736