refactor: restructure estimator to allow computing a subset of costs
The main idea here is to *not* return `RuntimeConfig` from `run` (which
requires us to compute all the costs, and fake out a bunch of unrelated
fields). Instead, we return a flat BTreeMap<String, Gas>, wrapped in a
nice API (`CostTable`).

We serialize the map into costs.txt file in a hand-rolled format
optimized for human readability and diffability. The actual config is
then restored out of this costs.txt file.

This is not yet implemented, but the idea is to make costs.txt a source
of truth for costs (we can just include_str! it into the binary).

While we do not implement computation of subset of costs in this PR
either, this now becomes possible -- a cost table can contain only a
subset of costs.

Finally, as a side bonus, this PR moves all actual logic for computing
the fees into the `run` method. While we do use `Measurements` for most
of the costs, we compute compilation, wasm instruction and trie node
touching costs in a special way.

The `costs.txt` file included in this PR is reconstructed from the
current RuntimeConfig.