Remove `PhantomData` from `Genesis` and always validate `Genesis` object on creation and deserialization. (#4592)
* Remove `PhantomData` from `Genesis`.
`Genesis` attempts to ensure consistency of `genesis.genesis_config.total_supply` with `genesis.records`, but in some cases this consistency can't be enforced. On top of that, the `Genesis` struct is `Serializable`, which means it's technically possible to construct via deserialization a struct with arbitrary values. Accept that and mark all fields as public, but warn the user to use the provided functions to instantiate and de-/serialize.

* Always validate Genesis, both on deserialization and creation.
It's possible to create a Genesis directly, but we assume people wouldn't do this.

* merge

* force validation upon Genesis construction, and set total_supply in test Genesis too

Co-authored-by: Sandi Fatic <sandi.fatic@gmail.com>