nayduck: add support for suffix in ./<path> include directive
Extend the `./<path>` directive so that it accepts arbitrary suffix
after the path.  If that suffix is present, all lines read from the
file are augmented by adding the prefix.  This makes it possible to do
things like:

    ./pytest.txt
    ./pytest.txt --features nightly_protocol,nightly_protocol_features

and avoid having to list every test twice in `pytest.txt` file.

Furthermore, when check_pytest.py and check_nightly.py read the files
also handle `#./<path>` so that a whole include can be commented out
with a TODO and the check scripts will treat files listed in those
files as being mentioned.