nearcore: return data from start_with_config in a struct rather than tuple
Rather than relying on positions in a tuple, change start_with_config
function to return its data in a structure so that the elements can be
named.  This will also make it easier to add new return values and
thanks to the use of ‘..’ callers won’t be affected.

Issue: https://github.com/near/nearcore/issues/4489