pytest: download `neard` binary i/o `near`
For now update the download URL to refer to the `neard` file rather
than `near` (they are the same binary; at some point `near` got renamed
to `neard`) only.  The file is still downloaded as `near-<branch>` and
there are other places in the code which expect it to be called that
way.