feat(neard): stop RPC servers gracefully after getting a signal (#4681)
With the default signal handlers actix_web HTTP servers uses, sending
a SIGINT results in a forceful shutdown of the servers even if there
are requests in flight.  This is unfriendly to the clients which may
as a result get a partial response or a broken connection.

Disable the signal handlers and instead explicitly call `stop(true)`
on the HTTP servers.  This gives the workers time to finish their
processing.  As a result, if a client successfully connected to an RPC
endpoint, they will get their response.  (That is, of course, unless
processing the request would take more than five seconds but that
should’t be the case with any of the supported RPC calls).

Fixes: https://github.com/near/nearcore/issues/4489