pytest: verify that `genesis-populate` run succeeded
Popen.wait does not check whether the executed command failed or not and
there was nothing in genesis_populate that would test that.  As a result,
the test would ignore any failures in the genesis-populate tool.  Change
the code to use subprocess.check_call instead which raises on failure.

Issue: https://github.com/near/nearcore/issues/4574