Rename `key_hash` from `TrieRefcountChange` to `trie_node_or_value_hash` (#4692)
As it was mentioned [here](https://github.com/near/nearcore/pull/4558#discussion_r676181684), `key_hash` is not a hash of trie key, but I missed some places while fixing it. Renamed it to ~~`value_hash`~~ `trie_node_or_value_hash`, because it is actually hash of the value.

Also, add some clarity on how the reference count is actually stored and updated.