Implement ShardUId (#4659)
This PR implements ShardUId as specified in near/NEPs#241 so that trie states are indexed by 'shard_uidinstead ofshard_id`.

The main change is to change the interface around trie storage to use shard_uid instead of shard_id. The database columns ColState and ColTrieChanges are now indexed by shard_uid instead of shard_id.

Changes that this PR does not include:

change account_id_to_shard_id to account for changing shard layout
gc
Test Plan:
The main source for errors is where shard_id and shard_uid is mixed up or the wrong shard_version is used for shard_uid, which may result in DB not found error. To test for the first type, we will change a wide range of existing unit and integration tests to use a shard layout with shard_version = 1 so shard_id and shard_uid cannot be mixed up.

Co-authored-by: Bowen Wang <bowenwang1996@users.noreply.github.com>