refactor(jsonrpc-primitives): deduplicate BlockReference (#4769)
The `BlockReference` enum is duplicated in `near-primitives` and `near-jsonrpc-primitives`. The likely reason for this, was to allow for jsonrpc-specific features [Context: <https://github.com/near/nearcore/pull/3861#discussion_r566134004>], but at this moment, they are, for the purposes that matter, identical. With the one in `near_jsonrpc_primitives` plainly convertible `Into` the one in `near_primitives`.