fix(primitives): backwards compatible ExecutionOutcomeView (#4851)
Since #4438, the `metadata` field was introduced on `ExecutionOutcomeView`.
Older nodes _(e.g the ones deployed on `[archival-]rpc.<network>.near.org` as of today)_ do not include that field. This means deserializing that data into the new `ExecutionOutcomeView` fails.

This PR introduces a default value for that field. Specifically, from `ExecutionMetadata::V1`.