Remove TestEnv::new_with_runtime constructor in favour of a builder
Replace all invocations of TestEnv::new_with_runtime constructor with
uses of the TestEnvBuilder.  This (hopefully) makes the code more
readable by naming the arguments and the use of the builder provides
a way for future expansion of the ‘constructor’ arguments.

This leaves TestEnv::new as the remaining constructor.  It will be
dealt with in a follow-up commit.

Issue: https://github.com/near/nearcore/issues/4835