runtime-tester: add support for specifying home directory
When debugging a test it may be desired to be able to see the state of
the storage after the tests finishes.  Currently, scenarios which use
on-disk storage will use a temporary directory which will be deleted
after the scenario runs.  This of course is correct behaviour but it
makes it impossible to investigate the storage of a failing (or
otherwise misbehaving) test.

Provide a Scenario::home_dir field which specifies where the store
will be located.  Importantly, Secnario::run will not delete that
directory after executing the scenario thus letting the user see
the end result.

Issue: https://github.com/near/nearcore/issues/4771