runtime-tester: add support for inspecting the home directory (#4868)
Currently, scenarios using on-disk storage use a temporary directory
which is deleted once the run finishes.  This is correct behaviour but
it makes it impossible to investigate the storage of a failing (or
otherwise misbehaving) test.

Change Scenario::run such that it returns the temporary directory
where the storage is held.  This way, caller can choose to hold onto
the directory and thus prevent it from being deleted.

Issue: https://github.com/near/nearcore/issues/4771