refactor: cleanup handling of compilation and caching errors in vm_runner
Before we run a wasm contract, we need to compile it to machine code.
There are two categories of errors that can happen during this process.

_First_, the contract might be just an invalid wasm. In that case, we
should record that as an error into the state.

_Second_, we might get an IO error when trying to use compiled contract
cache due to the database being corrupt. In this case, we should emit a
critical error and exit.

We used to conflate the two paths into returning just a
`Result<VMError>`, which additionally contains error which *can not*
happen at this phase (that is, function runtime errors).

So in this PR we change `Result<T, VMError>` into more precise and
reliable `Result<Result<T, CompilationError>, CacheError>`.

Concretely, this brings the following benefits:

* we no longer need to mark the whole of `VMError` as `Serialize`, as we
  only store compilation errors in the cache.
* we can not accidentally mix logical and fatal errors when working with
  compilation

The following I think is still an issue with code, to be fixed in some
future refactor:

* VM Error (logic error) contains a CacheError (fatal error) as a
  variant.

closes #4217
closes #4375