pytest: add testing protocol version to `upgradable.py` test
Extend the `upgradable.py` test (which is run from CI) to compare
protocol version in the stable and current releases.  The protocol
version is read from the `--version` screen of a `neard` binary.  This
way, the tests verifies the actual versions without having to manually
maintain a ‘current mainnet version’ constant somewhere.

Fixes: https://github.com/near/nearcore/issues/4956