feat: nightly feature to lower regular_op_cost (#4959)
The previous (safety multiplied) cost was 3_856_371 the new cost is
2_207_874.

The new cost was estimated using new parameter estimator

```
$ git clone git@github.com:matklad/nearcore.git
$ cd nearcore && git switch matklad/estimator-v2
$ git rev-parse --short HEAD
dece3552
$ cargo run --release -p runtime-params-estimator --features required -- \
    --docker --v2 --full \
    --vm-kind wasmer2 \
    --metrics-to-measure WasmInstruction
```

I've reproduced the measurments on local machine and on gcloud.

Curiously the raw (unmultiplied) results I get for wasmer0 and wasmer2
are

    WasmInstruction  1_566_236 gas  (computed in 12.57s)
    WasmInstruction    735_958 gas  (computed in 9.63s)

This is a bigger diff than what we've observed for actual costs. The
reason for that is that today's active cost is for some reason lower
than the measured cost. I don't know why that is the case.

Test Plan
------------

Not sure what's the best way to test this. This is a config change, so changing that literally this config has new value doesn't seem that useful. It would be nice to check that I didn't accidentally modify some other costs (i verified that manually with `diff`).