feat: update & enable jemalloc
Although we unconditionally depended on `jemallocator`, we didn't
actually set it as our global allocator (`#[global_allocator]` attribute
was missing, and the `cfg` condition was always false). So, we were
using systems allocator, which is usually slower than jemalloc, and adds
extra variability to our runtime (allocation behavior depdends on the
host OS version, and not on the neard version).

This PR fixes the configuration. It also substitutes tikv-jemallocator
for jemalloctor crate -- this is the maintained version of the crate and
something that, eg, rustc is using.

Test Plan
---------

cargo check -p neard && cargo check -p neard --no-default-features && cargo check -p neard --features performance_stats