try to enable simple nightshade in betanet again (#4990)
Yet another attempt to enabled simple nightshade in betanet. 

The issue last time was that the genesis file for betanet is regenerated at every deploy if the deploy changes protocol or db version. The `simple_nightshade_shard_layout` field is incorrectly serialized to be null in the newly generated genesis file and causes the simple nightshade upgrade process not to be triggered. 

Also adds a bunch of more debug infos.