pytest: add ‘spin_up_cluster’ sanity check to CI
Add a sanity test which simply spins up a cluster of nightly releases
to see if that’s still possible.

Add it to ‘upgradable’ step because it feels like a waste of resources
to create a new step for this trivial test.  `upgradable.py` test
builds the binary (albeit a non-nightly one) so hopefully we can share
some of the cache and stuff.

Fixes: https://github.com/near/nearcore/issues/4993