fix(test): prevent a race condition on neard start (#5007)
There is a potential race condition between when the peer manager actor starts and when the client actor starts. The issue is that if a peer establishes connection before client is initialized, `fetch_client_chain_info` won't get anything and as a result peers cannot be properly initialized and this caused some tests to fail. It is still not clear to me what change in nearcore caused this as it was not observed before, but this PR fixes the issue by always return *some* `ChainInfo` in `ViewClient`.

Fixes #4928 

Test plan
----------
http://nayduck.near.org/#/test/232592 @mina86 those scp failures :(