fix(chain): garbage collection should not cause node to crash (#5081)
In #5040 we introduced a change that could try to access already garbage collected information in the call to `get_shard_uids_to_gc`. More specifically, in `get_next_epoch_id_from_prev_block` we would try to access the block info for the first block of the epoch, which is presumably already garbage collected at this point. The reason why we did not catch it in tests is because epoch manager has a cache of size 1024 and we do not clean up the cache properly during garbage collection since `EpochManager` is not part of `Chain`.

Fixes #5074 

Test plan
----------
`cargo test -p integration-tests --features no_cache test_gc_long_epoch`