Move near_block_produced_total to chain/epoch_manager
Not that there's anything wrong with where it currently is, but it's
nice to increment the produced and expected metrics as close together
as possible. Before this change there's a large enough window that
when running pytests, the blocks produced metric is often higher than
blocks expected, which looks kindof weird.