runtime: report GasExceeded or GasLimitExceeded based on which limit is lower
If during execution of a contract both prepaid gas and max burnt gas
limits are exceeded, the `GasCounter` would report GasExceeded error.
This was purely based on which condition was checked first and
entirely accidental based on how the code was written.

Change the code such that both limits are tested and if both were
breached prefer reporting the one which is lower.  The edge case where
both limits are equal are handled in favour of reporting GasExceeded.