ci: add `performance_stats` to nightly CI tests
With cargo features comes the gubernatorial explosion of possible
versions of the code that need to be tested.  In current CI
configuration we are testing (i) code with no features and (ii) code
with nightly_protocol and nightly_protocol_features features enabled.

As it turns out, we run betanet nodes which enables performance_stats
feature which is variant of the code we are currently not testing.
Unfortunately, a bug in code enabled by that feature caused a betanet
failure.  To avoid further such issues, add performance_stats feature
to the nightly tests run.

The disadvantage of this change is of course that now variant of the
code with nightly features enabled but performance stats disabled is
not tested.  Hopefully though, since we are also testing code with no
features, overall this change increases coverage of code paths.