Ignore flunky tests (#5272)
Recently we added `test_blocks_ahead_of_head` and `test_timestamps`. Those tests are testing whenever a global variable `metrics` is updated properly. Unfortunately, we run multiple tests in parallel so it's not going to work.

We should disable those tests for now.

This is unrelated to this PR, but we should refactor `metrics` code to be readable like we it in `https://github.com/near/nearcore/pull/5265`.