refactor: use near_network_primitives directly (#5237)
Hey, this is going to be an epic change, but it will improve code quality.

We currently have a package called `near-network-utils`. For some reason we are important all types through `near-network`.
From not on we will import all packages directly.

Also `blacklist_from_iter` got moved out of `near-network`. It's not even used there.

The final goal is to remove dependency on `near-network` from packages that don't need it.