utils: use internally-mutable validated flag in MaybeValidated
Change MaybeValidated type from an enum with two variants (indicating
validated and not-validated payload) into a struct which holds
validated flag and payload in separate fields.  Furthermore, hold the
flag inside of a `Cell` to allow internal mutability.

This allows for an `validate_inplace_with` method which validates the
payload if necessary and updates the flag so that the next time the
method is called no validation needs to happen.

The downside is that MaybeValidated is no longer Sync or Send.

Issue: https://github.com/near/nearcore/issues/3863