do not include contract compilation time into wasm opcode estimation
The time it takes to compile the contract is completely orthogonal to
the time it takes to execute one instruction in  compiled contract.

Previously, as the cache was `None`, the code recompiled the contract
from scratch every time. Now it is only compiler during warmup
operation.

Using mock (in-memory) contract cache here is OK, as cache operations
are accounted for separatelly in the function call cost.