nfc: add a trait to abstract over a VM runtime
This is a non-functional-changes-intended refactoring. The primary goal
was to reduce the number of panicky branches in
`near-vm-runner::runner`. This is achieved primarily by two distinct
changes:

The runtime and the implementation of its `run` method has been moved
into the module which implements the runtime itself.

In addition to this, when creating the runtime, users are informed of
the absence of the runtime via an option which they can handle to
produce a more meaningful behaviour (e.g. tests could “simply” ignore
the VM kind) when a runtime is not available.

None of the callers should observe a change in the behaviour, as of this
change, however.

As a future cleanup, the `Runtime`s could start holding within them
certain data that's constant throughout the execution of the daemon,
such as `VMConfig`, `current_protocol_version` or
`CompiledContractCache` to reduce the amount of setup that needs happen
each time a contract execution is attempted.