bug: fix sync errors in `Clion` (#5403)
Clion prints lines about missing `from_bytes` due to the way we import `Signature`, This should fix it.

In this file we use ed25519_dalek::ed25519::signature::Signature struct method. Apparently, require you to import a type, if you use it's method with iit's trait. Even though it's name is not used directly. I'm sure @matklad can explain why it's done like this. Although we need to import it, it doesn't name which name we use for importing it.

So I can't just use alias Signature due to name collision, because there is already another Signature struct it in file,

As to why I'm renaming, _ to _Signature. The checked that's used in Clion has an issue, it complains about another error where there are multiple imports as _. It's a bug in that checked.

We could get rid of 'as _Signature' if we could rename the other struct from Signature to something else. Then importing signature::Signature as Signature would work.

@matklad It may be good to report to people who wrote it.