refactor: minimize work-around
`use Trait as _` is a standard Rust idiom -- it allows calling trait's
methods without bringing the trait itself in scope. This is useful to
avoid naming conflicts, which is exactly what happens in this case
(there's a `Signature` struct defined in this file).

We *could* `as _` the other two crates as well, but we don't as:

a) we don't have an active name conflict for them
b) more than `_` confuses CLion: https://github.com/intellij-rust/intellij-rust/issues/8143

Notes:

Rust naming convention suggests using verbs as trait names, which helps
with dodging similar conflicts.