Improve the error messages from prepare.rs
Contract preparation is the first code that will typically see the
contract code and consequentially it is also the first which will report
issues with the contract.

One such example is #5310 for which the preparation code should now
report something more meaningful.

I went ahead and implemented `std::error::Error` in some places, but it
seems to me like we probably want to move the entire file to
`#[thiserror]`...