Verify the CPU `neard` runs on more meticulously
This changeset does a couple different things. First and foremost it
makes `neard` verify the environment it is running in. It'll do so only
for production builds by default, making regular `cargo run` or `cargo
build` workflows work as usual.

However in production artifacts (as produced by `make release`, the
`cfg(near_production)` will be set, various CPU features will be enabled
for LLVM codegen purposes and the `neard` itself will verify that the
prerequisite features are available before it does anything else.

Note however, that there is no guarantee that the code, which runs before
`main` is invoked, avoids use of these potentially unsupported features,
possibly leading to less than stellar failure modes on machines which
wouldn't be able to run `neard` either way.

This may have an effect on gas costs for host functions, and for that
reason this change must not be merged before we quantify those changes.
As a less obvious outcome, this may also cause cost estimations done on
the production build differ from cost estimations done with a regular
`cargo built` code.