scripts: add a `check_formatting.sh` script
Introduce a new `script/check_formatting.sh` script which incorporates
Python formatting check currently coded in the BuildKite pipeline
command directly and Cargo formatting check done as an integration
test.

This has two benefits.  For the Python code, the pipeline command is
now less complex.  Call to yapf and an error message on failure are
now replaced by a `check_formatting.sh` invocation.

For the Rust code, the current situation is such that if code is not
formatted correctly ‘cargo test’ as well as ‘cargo test nightly’ steps
are failing.  With this change there’s going to be just one failure in
‘sanity checks’.