pytest: make A/B tests against latest deployed mainnet binary
Use latest_release and latest_deploy data stored on S3 rather than
relying on GitHub releases to figure out what is the latest binary run
on mainnet for performing A/B tests.

This fixes issues around release time when our CI suddenly breaks when
a new rc is made without it yet being released on either testnet or
mainnet.

For example, once 1.42.0-rc1 is release on GitHub, the code expects to
download a 1.42.0 binary from S3 regardless of whether such release
has been deployed or not.  With this change, the code will look at the
actual executable that is running in production on mainnet.

This also makes sure that all our tests are performed against what is
running on mainnet so that we can confirm that upgrading from mainnet
to current HEAD is possible.

Fixes: https://github.com/near/nearcore/issues/4956