Revert "Revert "Request chunks for multiple blocks simultaneously (#5… (#5637)
…224)" (#5473)"

This comment commits the change for "Request chunks for multiple blocks simultaneously again".

Initially I reverted the change because testnet canary ran into OOM issues and we suspect that change caused the issue. However, I couldn't reproduce the issue when I started a new testnet rpc node (I restarted the node several times to test when it was behind). I am going to revert this commit because 1) testnet release is finished 2) it's good to test the change in betanet before the next testnet release

I also discovered a few other issues in the current implementation regarding requesting chunks, some of them will help with reducing network requests if that was the culprit. I will address them in separate PRs. After this, I will work with @nikurt to start a mocknet testing all these changes together.