fix: correctly include WASM instruction cost into profile
Our gas profiling was incomplete, as it didn't include the cost for wasm
instructions. One would thing that it is enough to subtract the sum of
all other costs from the total to get one, but that's not entirely
correct -- the total gas_burnt the user sees in the RPC includes various
fees computed outside of near_vm_runner. So let's explicitly account for
WASM execution costs in profile (luckily, adding new costs is a
backwrads-compatible change).

While we are at it, let's also add an integration tests that asserts
that this time we indeed accounted for all of the costs.