feat: Add bandwidth stats to `PeerManagerActor` pt2 (#5629)
Continuation of https://github.com/near/nearcore/pull/5455

We are in progress of determining, what are the safe ranges of limit thresholds to set for `ThrottleController`.

`num_messages_in_progress` is part of `ThrottleController`. It's associated with given `PeerActor`. I tracks the number of messages which are currently in progress of being processed.

I added stats, which help us track, what was the maximum value of `num_messages_in_progress` across all peers over period of 1m. This will help us setup the liimits.