chore: upgrade rust-openssl
This makes sure that we use the latest version as of the time of
writing.

Notably, both openssl-sys 0.9.67 an 0.9.72 default to using openssl
1.1.1. openssl-sys 0.9.71 used openssl 3.0.0, and that was observed to
cause problems, so upstream reverted to 1.1.1, which is still supported.

There's no direct benefit to the PR, it just makes sure that we track
the latest upstream version.

closes #5109