refactor: don't abuse serialization for type erasure
vm-logic is "generic" over External implementation. We don't want
however to literally add a type parameter to it, so we use `dyn
External`. This creates a problem that, when a particular external
fails, we need to report the error, but we don't know error's concrete
types. So let's use `Box<dyn Any>` for that, and a bit of down-casting.

The previous impl did roughly equivalent thing, but via
serialize/deserialize route, which is more complicated than what we need
here.

As a result, we eliminate a couple of BorshSerializes and clones.