scripts: add `check_fuzzing.py` checking if fuzz targets are listed
Introduce `scripts/check_fuzzing.py` file which scans the repository
to find all defined cargo-fuzz targets and sees if all of them are
referenced in nightly tests set.  This serves the same purpose as
`check_pytest.py` and `check_nightly.py` scripts which verify that
all pytests and expensive tests respectively are referenced.