chain: indexer: avoid allocation of temporary vector
In the two places where `fetch_chunks` function is called, it is used
to fetch chunks for a block.  By moving code which extracts list of
chunk hashes from a block view into the `fetch_chunks` function we
reduce code duplication and eliminate an allocation of a temporary
vector object.

To better describe what the function does, rename it to
`fetch_block_chunks`.