feat(rpc): add chunk production stats to the /validators method (#5843)
* feat(rpc): add chunk production stats to the /validators method

It's a small change to add this info, as it's already present in
the EpochSummary and EpochInfoAggregator structs. This should not
be a breaking change (after removing Borsh derives), since fields are only being added.

Test plan
---------
Since this isn't adding much new logic, just exposing the result of
existing logic, it should be sufficient to just manually check that
the output looks correct/in line with the current chain

Issue: https://github.com/near/nearcore/issues/5708

* remove BorshSerialize/BorshDeserialize derives from CurrentEpochValidatorInfo

These aren't used anywhere, and their usage could make adding fields a
breaking change

* make sure to mark new CurrentEpochValidatorInfo fields with #[serde(default)]

Otherwise the absence of these fields (when connecting to an old client)
would cause a parse failure