fix: don't log host functions by default
We received feedback that near_vm_runner produces too much output couple
of times. At the same time, there was once instance where having a log
line for something as trivial `read_register` was useful.

So, let's downgrade the logging here to trace for now, and maybe revisit
in the future when we have a clearer understanding of which amount of
logs is acceptable.