refactor: use the lower level wasmer APIs
This is intended to be a non-functional change.

Here we replace the use of the `wasmer` crate and the wrappers it
provides with the lower level APIs exposed by various wasmer subcrates
such as `wasmer-engine`, `wasmer-vm`, etc.

These changes are primarily intended to reduce the burden of maintaining
and publishing the `wasmer-near` crate. Since our requirements for a
runtime are scoped to very narrow portion of what the `wasmer` crate
provides, the complexity of the code in `nearcore` didn't
change as much as I feared it would, while also making certain portions
of the codebase significantly more straightforward.

The `imports`-related code is perhaps the highest gain in complexity of
the code, as it has become necessary to take care of ensuring a matching
ABI. On the other hand, the changes there also reduced non-local
assumptions about the lifetime of the `VMLogic` by moving it entirely to
the funcition which handles instantiations.

There are definitely more opportunities for simplifying the code around
this part of the codebase, but these will need to happen as a follow up
as this one is already getting quite large, and some of the changes
would touch other runners, not just Wasmer2.