Makefile: add `neard-sandbox` executable name
For consistency with plain release naming, we probably should prefer
`neard-sandbox` to `near-sandbox` executable name.  This commit
changes `sandbox` and `sandbox-release` make targets to provide this
new name.

This does not remove the old name since God knows how many things
depend on it (we are uploading sandbox binary as `near-sandbox` and
documentation refers to that name) but having the new name will allow
us to phase out the legacy one with time.

Issue: https://github.com/near/nearcore/issues/5969