fix(chain): fix block ordinal computation (#5761)
In #4793 we unfortunately did not fix the issue fully. The situation that we did not fix is as follows: if block h is a fork and block h+1 is on the canonical chain, they would have the same ordinal and if block h is received later than block h+1, it would overwrite block h+1 in `ColBlockOrdinal`, which is incorrect.

This PR fixes the issue completely by updating `ColBlockOrdinal` in a similar way to how `ColBlockHeight` is updated. Essentially we only update it when the header head change. When the header head changes, we overwrite the ordinals between the last header height and the new header head. This way we guarantee that only hashes of blocks on the canonical chain eventually stay in the column given how the fork choice rule works.

Test plan
--------------------------
`test_block_merkle_proof`