refactor: make sure there's only one code path an InconsistentStateError can take through runtime (#5837)
This started as pure refactor to remove duplicate logic, but it uncovered a surprising behavior, at least in theory. It seems today runtime can upgrade `TreNodeMissing` error into `InconsistentStateError`, is this expected? 

Builds on top of #5813 and can be viewed per commit. The actuall change is the last commit, the stuff before are pure refactors which make the change more obvious. 

Previously, there were two different code paths an
StorageError could take:

Path 1:
* RuntimeExt emits StorageError
* RuntimeExt calls wrap_storage_error to wrap it into VMLogicError::ExternalError
* near_vm_runner translates that into VMError::ExternalError
* actions.rs matches on AnyError and emits a StorageError

Path 2
* RuntimeExt emits StorageError
* RuntimeExt wraps that into VMLogicError::InconsistentStateError. Note
  that this state is especially weird: `InconsistentStateError` is a
  *subset* of `StorageError`. But here RuntimeExt treats *any* storage
  error as InconsistentStateError.
* near_vm_runner translates that into VMError::InconsistentStateError
* actions.rs translates that into `StorageError::InconsistentStateError`

Note in particular how on the path 2 we can start with
`StorageError::TrieNodeMissing` and end with
`StorageError::StorageInconsistentState("TrieNodeMissing")`, which feels
like a bug maybe?

Turns out, that the second path is actually dead code, so we can safely remove it, what we actually do in this PR.