refactor: Runtime structure now contains VMConfig
Conceptually this matches what is happening somewhat better – VM is
constructed as configured in some specific way and then can be used to
run multiple contracts or somesuch, as long as the configuration
matches.

This kind of construction can help, in the future, to reuse more state
between contract invocations. Since `VMConfig` doesn't change within the
same run of `neard` typically, we could reuse the same memory map to
hold the contract code and cache the allocation of the working memory as
well (so long as we clear it between contract invocations), thus saving
on page faults and similar cost centres.