Manually set the target for wasmer Singlepass
QEMU with the TCG plugin does not support running AVX instructions and
makes it impossible to set the CPUID flag. This fails wasmer's CPU
feature checks when the default (host's) Target is used.

We already have a work-around for this kind of issue in place – the
`no_cpu_compatibility_checks` feature flag. Here we manually construct a
known-good target when this flag is enabled so that the param estimator
can run (so long as it does not actually invoke any NaN
canonicalization…)