neard: state explicitly that unsafe_reset_all deletes directory
The documentation of unsafe_reset_all subcommand says that it removes
all the config, keys and data but in reality it does more by deleting
the entire home directory and any and all files that reside inside of
it (which includes aforementioned files).  Mention that explicitly in
the documentation.

Issue: https://github.com/near/nearcore/issues/6010