cli: better respect RUST_LOG (#6036)
At the moment, when configuring what data is logged we first
initialise EnvFilter with a predefined defaults and only later if
RUST_LOG environment variable is set add that on top of the.  This
means that setting RUST_LOG to `warn` doesn’t actually filter messages
to include warnings and up only.

Change the code such that if RUST_LOG environment variable is set
(even if empty), the value is taken as is without including the
defaults.  If the variable is not present, the defaults are used as
before.

Fixes: https://github.com/near/nearcore/issues/6008