tests: fix flaky integration test
the download_file() test might fail if download_file() runs before
the HTTP server gets a chance to start. Fix it by waiting until the server
is already listening to call download_file(). This requires moving the
tokio runtime block_on() call to a wrapper function so we don't have problems
starting a runtime inside a runtime in the test code