chain: replace O(N) with O(1) hash lookup in assign_with_possible_repeats
To avoid adding chunk producer to a shard multiple times, the code
would go through the list of already assigned producers and look for
the one being considered.  This means a O(N) lookup each time so
overall a O(N²) algorithm.

Replace the linear scan with a hash set offering O(1) lookup which
reduces complexity of the algorithm.