scripts: run_docker.sh improvements and fixes (#6104)
Rather than `CMD` use `ENTRYPOINT`.  Using the image on Kubernetes
this allows the use of `args` field for additional parameters such as
adding `--disable-rpc` or any other flag.  Those options are passed to
the script so use `"$@"` to access them.

With that, use `exec` to run the daemon.  This is to not let bash as
PID 1 which is a best practice for signal and zombie handling, see
<https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#entrypoint>.

Finally, fix typos when checking if INIT and NODE_KEY environment
variables are non-empty, quote all the variables and don’t use any
bash features so that we can use `#!/bin/sh`.