re: Fix Clion false error in `near_network.rs` (#6076)
Clion Rust plugin for Intellij gives a false error. Let's fix it. Maybe we should inform the rust clion plugin team?

```
mismatches types [E308]
expected `&HandshakeAutoDes`, for `&mut &[u8]`.
```

This issue is not too important to fix. I reported the issue to `intelij-rust`, and this PR can be helpful in fixing the issue in the plugin.