mem_grow_cost = mem_grow_cost * regular_op_cost
In the past we'd charge for `memory.grow` using the same `gas` function
which would multiply the provided `memory.grow` cost with the
`regular_op_cost`, much like all the other calls to `env.gas`.

Now that we're premultiplying at instrumentation time, we must account
for `memory.grow` much the same way.