chain: rosetta: guarantee order of operations in genesis block
Because we were using `buffer_unordered` as well as a hash map the
order in which we consider accounts on genesis block is
nondeterministic.  This means that operations in the ‘genesis
transaction’ are in nondeterministic order as well.  This perhaps
wouldn’t be an issue except that we assign operation indexes to each
operation based on their position in the list.

This means, that querying genesis block multiple times may result in
transactions whose operation indexing changes.

To make sure that doesn’t happen and operation indexes remain
consistent, switch from using a hash map to a B-tree map.

Note that this fixes the issue for genesis block only.  Other blocks
still suffer from the ordering being nondeterministic.