Update to wasmer 2.2.0
This pulls in a fair amount of development that has occurred in the fork
over the past couple weeks.

Most notably it has a fix for a crash, functionality to simplify
profiling of the code generated by JIT, a stack limiting safeguard, and
a removal of superfluous CPU capability checks, allowing cost estimation
work again.