fix: Fix memory leak in `near-network` affecting `indexer` and probably all nodes (#5902)
There is a huge memory leak, which is affecting indexer and `near-network`.

After rewriting `memory-analyzer` from `C++` to `Rust`. I noticed that, there is a memory leak in `Codec`. The codec is not being cleared after it's length drops to 0.

I added a logic for clearing both `read` and `write` buffers. This should help indexer, and regular nodes as well

```shell
> sudo rust-memory-analyzer analyze --pid `pidof indexer-example` --print-raw-symbols 2>&1

 INFO memory_analyzer::analyze: func="_ZN5bytes9bytes_mut8BytesMut13reserve_inner17h599b80e87834931bE" count=52 size_mb=993
 INFO memory_analyzer::analyze: took=1.116597038s total_size_mb=1774
 INFO memory_analyzer::analyze: took=1.116604928s resident_but_not_used_mb=144 allocated_with_proxy_mb=1055 mapped_files_mb=575
```

After having indexer run for 10m, it was using almost `2gb` before, now it's down to `900mb`. That's over 1gb of memory saving.