fix: reduce number of chunk requesting messages (#6148)
This is a quick fix to reduce the number of chunk requests that we send. It changes the number of orphan blocks for which we requests for chunks from 5 to 2. 2 means we only request for missing chunks for at most 2 blocks if there are no forks. We should get most of the benefit from requesting chunks early with the number 2, since chunk requesting time is ~200-400ms, less than the block processing time these days.

Test plan:
- Start a node on testnet with the change and verify that it can catch up
- I restarted the node three times and each time it caught up every fast, CPU and memory usage was normal. I also verified in the log that there were indeed cases when the node requested chunks for two blocks at a time.
- https://grafana.near.org/d/zzGDUdcnk/chunks-delay-monitoring?orgId=1&var-chain_id=testnet&var-role=min-testing&var-node_id=All&from=now-3h&to=now