pytest: fix a failure in skip_epoch.py
In skip_epoch.py, there are two periods where a block producer
responsible for some shard is down. The first is when a validator
never went online to begin with, and the second is when we kill one
after the chain has been running for a while. In each case, we wait
for a couple epochs to pass and then check that the balances resulting
from a bunch of transactions are applied correctly, the idea being
that we want to check that some online validator is made responsible
for the offline shard and is able to sync its state.

In the second case, we record the latest block height before killing
the validator and then wait for EPOCH_LENGTH * 3 - 1 blocks, so that
we land at least in epoch T+2. But this may not be a long enough wait
period. If the validator had already produced some blocks in the epoch
in which it's killed, it might not actually drop below the kickout threshold
for that epoch, in which case we won't expect to see "height_included" in that
validator's shard move up to HEAD until epoch T+3.

So fix it by waiting for that long, and actually make it explicit by
introducing a utils.poll_epochs() function that does basically the same
thing as poll_blocks(), but for epochs instead