client: add GetBlockHeader request and use it in Rosetta RPC
The /network/status Rosetta RPC end point returns, among other things,
height and hash of the genesis block.  Currently, it does so by
fetching the genesis block through the view client but in non-archival
nodes that is not available and results in an error.  To address that,
introduce a GetBlockHeader message which returns only the header.
Since headers are not garbage collected, this request will always work
even for garbage collected blocks.