Enable rosetta_rpc feature by default
As we’re starting to test Rosetta RPC more and offering a public test
instances with Rosetta enabled, enable the support by default so we
don’t have to deal with multiple versions of binaries (some with and
some without support for Rosetta).

Note that the implementation needs to be activated explicitly at run
time so enabling the feature does not automatically mean the node will
have Rosetta end thread enabled.